% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_panel_function_helpers.R
\name{baselineGrob}
\alias{baselineGrob}
\title{Baseline grob}
\usage{
baselineGrob(coords)
}
\arguments{
\item{coords}{the transformed coordinates and stats}
}
\description{
This function allows you to create the baseline segmentGrob
}
