% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{binBounds}
\alias{binBounds}
\title{Calculate bin bounds}
\usage{
binBounds(value, bin.width)
}
\arguments{
\item{value}{the data you want to calculate the stats of}

\item{bin.width}{the bin width that will be used to plot the data}
}
\description{
This function allows you to calculate the upper and lower bounds of the bins
that include data (different to the maximum and minimum values)
}
