% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_panel_function_helpers.R
\name{confidenceIntGrob}
\alias{confidenceIntGrob}
\title{Confidence interval grob}
\usage{
confidenceIntGrob(coords, upper.or.lower)
}
\arguments{
\item{coords}{the transformed coordinates and stats}

\item{upper.or.lower}{whether it is the upper or lower bound being plotted}
}
\description{
This function allows you to create one of the linesGrob for the confidence
interval ticks
}
