% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{dfStats}
\alias{dfStats}
\title{Calculate stats of the dataset}
\usage{
dfStats(value, bin.width)
}
\arguments{
\item{value}{the data you want to calculate the stats of}

\item{bin.width}{the bin width that will be used to plot the data}
}
\description{
This function allows you to calculate all of the stats required for seastack
plots
}
