% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_panel_function.R
\name{draw_panel_function}
\alias{draw_panel_function}
\title{draw_panel_function}
\usage{
draw_panel_function(
  data,
  panel_params,
  coord,
  bin.width,
  mean.size,
  median.size,
  show.standard.dev,
  show.confidence.int,
  show.mean,
  show.median
)
}
\arguments{
\item{data}{the data being plotted}

\item{panel_params}{the parameters of the plot panel}

\item{coord}{the coordinates of the plot (I don't fully understand this)}

\item{bin.width}{the bin width being used to plot}

\item{mean.size}{the size of the mean diamond}

\item{median.size}{the size of the median point}

\item{show.standard.dev}{whether the standard deviation rectangle should be
plotted}

\item{show.confidence.int}{whether the confidence interval should be plotted}

\item{show.mean}{whether the mean should be plotted}

\item{show.median}{whether the median should be plotted}
}
\description{
draw_panel_function
}
