% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{histBreaks}
\alias{histBreaks}
\title{Calculate bin breaks}
\usage{
histBreaks(value, bin.width)
}
\arguments{
\item{value}{the data you want to calculate the stats of}

\item{bin.width}{the bin width that will be used to plot the data}
}
\description{
This function allows you to calculate the breaks between bins
}
