% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seastackplot.R
\name{sea_stack_plot}
\alias{sea_stack_plot}
\title{seastackplot}
\usage{
sea_stack_plot(
  data,
  data.column = "value",
  group.column = "group",
  data.label = NULL,
  bin.width = 1,
  hist.colours,
  mean.size = 4,
  median.size = 3.2,
  orientation = "vertical",
  show.mean = T,
  show.median = F,
  show.standard.dev = T,
  show.confidence.int = F
)
}
\arguments{
\item{data}{the data to be plotted}

\item{data.column}{the name of the column containing the data to be plotted}

\item{group.column}{the name of the column containing the group names for the data}

\item{data.label}{the desired title for the data axis, if left NULL (default), will be the value of data.column}

\item{bin.width}{the bin width being used to plot}

\item{mean.size}{the size of the mean diamond}

\item{median.size}{the size of the median point}

\item{orientation}{"vertical" or "horizontal" whether or not the plot is being
plotted vertically (default) or horizontally}

\item{show.mean}{whether the mean should be plotted}

\item{show.median}{whether the median should be plotted}

\item{show.standard.dev}{whether the standard deviation rectangle should be
plotted}

\item{show.confidence.int}{whether the confidence interval should be plotted}
}
\description{
Plot a sea stack plot with summary statistics and histogram! This is the
easiest way to plot a sea stack plot.
}
