% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{sem}
\alias{sem}
\title{Standard error of the mean}
\usage{
sem(value, na.rm = F)
}
\arguments{
\item{value}{the data you want to calculate the sem of}

\item{na.rm}{removes any NAs from value}
}
\description{
This function allows you to calculate the standard error of the mean (sem)
}
