% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_panel_function_helpers.R
\name{standardDevGrob}
\alias{standardDevGrob}
\title{Standard deviation grob}
\usage{
standardDevGrob(coords, vertical = T)
}
\arguments{
\item{coords}{the transformed coordinates and stats}

\item{vertical}{whether or not the plot is being plotter vertically (as
opposed to horizontally)}
}
\description{
This function allows you to create the standard deviation rectGrob
}
