% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_panel_function_helpers.R
\name{transformStats}
\alias{transformStats}
\title{Transform stats}
\usage{
transformStats(transformed.coords, df.stats, panel_params)
}
\arguments{
\item{transformed.coords}{transformed base coordinates}

\item{df.stats}{the calsulated statistics of the raw data (calculated by
dfStats)}

\item{panel_params}{the parameters of the plot panel}
}
\description{
This function allows you to transform the stats into the ggplot graph's
coordinate system
}
