% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_panel_function_helpers.R
\name{meanGrob}
\alias{meanGrob}
\title{Mean grob}
\usage{
meanGrob(coords, mean.size = 4)
}
\arguments{
\item{coords}{the transformed coordinates and stats}

\item{mean.size}{the size you want the mean point to be}
}
\description{
This function allows you to create the mean pointsGrob
}
