% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_panel_function_helpers.R
\name{medianGrob}
\alias{medianGrob}
\title{Median grob}
\usage{
medianGrob(coords, median.size = 3.2)
}
\arguments{
\item{coords}{the transformed coordinates and stats}

\item{median.size}{the size you want the mean point to be}
}
\description{
This function allows you to create the median pointsGrob
}
