% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_panel_function_helpers.R
\name{standardDevGrob}
\alias{standardDevGrob}
\title{Standard deviation grob}
\usage{
standardDevGrob(coords)
}
\arguments{
\item{coords}{the transformed coordinates and stats}
}
\description{
This function allows you to create the standard deviation rectGrob
}
